Fix su handling of --
If getopt_long runs across a -- , it will leave you at the next
argument, not at --.  But if you have - before --, then it will
leave you at the -.  So, after getopt_long,

First check whether previous argument is --.  If so, record it
so that we don't parse next args as username.

After checking for '-', check whether -- is next.  If so, then
record that so that we don't parse next args as username.

Fix an off by one when rewinding the argv pointer.  We were skipping
the first argument.

Closes #253

Signed-off-by: Serge Hallyn <shallyn@cisco.com>