ustr2str.c: Add function to make strings from unterminated strings
The implementation is similar to strncpy_s(3), but avoids the
unnecessary runtime checks.  It's better to wrap the function in a
macro and do as many static_asserts(3) as one considers necessary;
In fact, if in the future C allows backwards VLA syntax, static
analysis could be better than those static_assert(3)s.

Also, this function enforces that (proper) strings and unterminated
strings are not mixed by accident, since it uses distinct types for
them.

Also, the name of the function clearly documents its functionality,
unlike strncpy_s(), which looks like it has something in common with
the dreaded strncpy(3).

The design of this function was improved with feedback from Andrew
Clayton <a.clayton@nginx.com>.

Cc: Andrew Clayton <andrew@digital-domain.net>
Signed-off-by: Alejandro Colomar <alx@kernel.org>