commonio: Add fgetsnulx() and getlinex()
fgets(3) has a useless return value.  Instead, it would be better if it
returned a pointer to the terminating NUL byte.  For our fgetsx(), let's
write such a function.  It will help implement getlinex().

getline() is a complex function, and open-coding it is error-prone.  A
variant that handles nicely newline escapes is even more complex.  Let's
write getlinex(), and use it instead of open-coding it.

Signed-off-by: Alejandro Colomar <alx@kernel.org>